avoid warning.
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@1337 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ x11/gtk2/gtk_menu.c
@@ -1722,7 +1722,7 @@ cb_soundboard(gint idx)
 static void
 cb_radio(GtkRadioAction *action, GtkRadioAction *current, gpointer user_data)
 {
-	guint value = (guint)gtk_radio_action_get_current_value(action);
+	gint value = gtk_radio_action_get_current_value(action);
 	guint menu_idx = (guint)GPOINTER_TO_INT(user_data);
 	gint i;
